+++ Changelog.md
@@ -1,8 +1,13 @@
 #Changelog      
 
+##1.2      
+- Added Config File      
+- Added Blacklist for bedification      
+- Added Hardcore Mode, replaces everything with bedrock      
+
 ##1.1
-- Remove bed placement on end portal frames
+- Remove bed placement on end portal frames      
 
 ##1.0
 - Initial Release
-- **Everything is a fricken bed!**
+- **Everything is a fricken bed!**